Fixed tests and references
