+++ python/kfserving/kfserving/__init__.py
@@ -13,9 +13,12 @@
 # limitations under the License.
 from __future__ import absolute_import
 
-from .server import KFServer
-from .model import KFModel
-from .storage import Storage
+from .kfmodel import KFModel
+from .kfserver import KFServer
+from .storage import storage
+from .constants import constants
+from .utils import utils
+from .handlers import http
 
 # Below is merged from kfserving client.
 # import ApiClient
@@ -23,10 +26,7 @@ from .api_client import ApiClient
 from .configuration import Configuration
 # import client apis into kfserving package
 from .api.kf_serving_client import KFServingClient
-# import constants into kfserving package
-from .constants import constants
-# import util into sdk package
-from .utils import utils
+
 # import models into kfserving package
 from .models.knative_condition import KnativeCondition
 from .models.knative_volatile_time import KnativeVolatileTime
