+++ python/kfserving/kfserving/handlers/http.py
@@ -1,12 +1,12 @@
-from http import HTTPStatus
 import tornado.web
 import json
-from kfserving.model import KFModel
+from typing import List, Dict
+from http import HTTPStatus
+from kfserving.kfmodel import KFModel
 
 
-class V1Handler(tornado.web.RequestHandler):
+class HTTPHandler(tornado.web.RequestHandler):
     def initialize(self, models: Dict[str, KFModel]):
-        self.protocol = protocol
         self.models = models
 
     def get_model(self, name: str):
@@ -34,10 +34,29 @@ class V1Handler(tornado.web.RequestHandler):
                 status_code=HTTPStatus.BAD_REQUEST,
                 reason="Expected key \"instances\" in request body"
             )
-        
+
         if not isinstance(body["instances"], list):
             raise tornado.web.HTTPError(
                 status_code=HTTPStatus.BAD_REQUEST,
                 reason="Expected \"instances\" to be a list"
             )
 
+
+class PredictHandler(HTTPHandler):
+    def initialize(self, models: Dict[str, KFModel]):
+        self.models = models
+
+    def post(self, name: str):
+        model = self.get_model(name)
+        self.validate(model)
+        self.write(model.predict(self.request.body))
+
+
+class ExplainHandler(HTTPHandler):
+    def initialize(self, models: Dict[str, KFModel]):
+        self.models = models
+
+    def post(self, name: str):
+        model = self.get_model(name)
+        self.validate(model)
+        self.write(model.explain(self.request.body))
